// packages/unlighthouse-old-mono/src/types.ts
export interface RouteDefinition {
  name: string
  path: string
  file?: string
  component?: string
  componentBaseName?: string
  chunkName?: string
  _name?: string
  layout?: string
}

/**
 * A normalised route, in the context of unlighthouse, is a mapping between a URL path, it's definition and a unique id.
 */
export interface NormalisedRoute {
  id: string
  path: string
  url: string
  $url: URL
  definition: RouteDefinition
  /**
   * A runtime path that the route was discovered from, useful if the route is a 404 and we want to know what directed
   * us to it.
   */
  discoveredFrom?: string
}

export interface ComputedLighthouseReportAudit {
  details?: {
    items?: any[]
  }
  displayValue: string | number
  score: number
}
/**
 * An augmented Lighthouse Report type, we add custom types to the base report for specific functionality on the
 * @unlighthouse/client.
 */
export type LighthouseReport = Partial<Result> & {
  /**
   * The total score for the result, this is the sum of each category's result
   */
  score: number
  categories: { score: number | null }[]
  computed: {
    /**
     * An aggregation of multiple image audit results.
     */
    imageIssues: ComputedLighthouseReportAudit
    ariaIssues: ComputedLighthouseReportAudit
  }
}

/**
 * Tasks that Unlighthouse will run, used to track their status.
 */
export type UnlighthouseTask = 'inspectHtmlTask' | 'runLighthouseTask'

/**
 * Each task ran by unlighthouse (extractHtmlPayload, runLighthouseTask) has a specific status which we can expose.
 */
export type UnlighthouseTaskStatus = 'waiting' | 'in-progress' | 'completed' | 'failed' | 'ignore' | 'failed-retry'

/**
 * A fairly rigid representation of the puppeteer cluster task results (extractHtmlPayload, runLighthouseTask), combined
 * with the normalised route.
 */
export interface UnlighthouseRouteReport {
  /**
   * The mapping of tasks with their status.
   */
  tasks: Record<UnlighthouseTask, UnlighthouseTaskStatus>
  /**
   * Track how long tasks are taking
   */
  tasksTime?: Record<Partial<UnlighthouseTask>, number>
  /**
   * Path to where the artifacts from a URL scan are saved.
   */
  artifactPath: string
  /**
   * URL of where the artifacts are stored, for static client access.
   */
  artifactUrl: string
  /**
   * The route (URL Path) that the report belongs to.
   */
  route: NormalisedRoute
  /**
   * A unique representation of the route, useful for the API layer.
   */
  reportId: string
  /**
   * The lighthouse result, only set once the task is completed.
   */
  report?: LighthouseReport
  /**
   * The SEO meta-data, only set once the html payload has been extracted and passed.
   */
  seo?: HTMLExtractPayload
}

export interface HTMLExtractPayload {
  alternativeLangDefault?: string
  alternativeLangDefaultHtml?: string
  title?: string
  description?: string
  internalLinks?: number
  externalLinks?: number
  htmlSize?: number
  favicon?: string
  og?: {
    description?: string
    title?: string
    image?: string
  }
}